quantum-guard-backend/quantum-guard-backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
python-multipart
bcrypt
pyjwt
python-dotenv
pytest
httpx